requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
pyyaml>=6.0
requests>=2.28.0
